# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas API Contract
  version: 1.0.0
  summary: Master API contract for Atlas platform clients and services.
  description: |
    Production-grade OpenAPI 3.1 contract for Atlas.

    This specification is the single source of truth for:
    - Web frontend
    - Backend implementation
    - Mobile clients
    - Admin portal
    - External integrations

    Design guarantees:
    - RESTful, versioned API under /api/v1
    - Consistent response envelope
    - Standardized errors
    - Standardized pagination/filter/sort/search
    - Typed request and response contracts

    Note: This contract defines interfaces only. It intentionally contains no implementation logic.
  contact:
    name: Atlas API Platform Team
    email: api-platform@atlas.local
  license:
    name: Proprietary
...
servers:
  - url: https://api.atlasbank.com/api/v1
    description: Production
  - url: https://staging-api.atlasbank.com/api/v1
    description: Staging
  - url: http://localhost:3001/api/v1
    description: Local development
...
security:
  - bearerAuth: []
  - sessionCookie: []
paths:
  /auth/register:
    post:
      tags: [Authentication]
      summary: Register a customer account
      description: |
        Creates a new customer account and initial profile shell.

        Validation:
        - Email must be RFC-compliant and unique.
        - Password must satisfy platform password policy.
        - Terms and privacy acceptance timestamps required.

        Permissions: Public
        Rate limit: 10 requests / 15 minutes per IP
      operationId: registerCustomer
      security: []
      x-permissions:
        - public
      x-rateLimit:
        scope: ip
        burst: 10
        window: 15m
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            examples:
              default:
                value:
                  email: ava.simmons@example.com
                  password: StrongPass!234
                  firstName: Ava
                  lastName: Simmons
                  termsAcceptedAt: '2026-07-09T13:05:00Z'
                  privacyAcceptedAt: '2026-07-09T13:05:00Z'
      responses:
        '201':
          description: Customer account created.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthSession'
              examples:
                created:
                  value:
                    success: true
                    data:
                      accessToken: eyJhbGciOi...
                      refreshToken: eyJhbGciOi...
                      tokenType: Bearer
                      expiresIn: 900
                      sessionId: sess_7a2f31
                    meta:
                      traceId: trc_1001
                      timestamp: '2026-07-09T13:05:04Z'
                    error: null
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate customer credentials
      description: |
        Authenticates a customer and returns an access/refresh token pair.

        Validation:
        - Email and password required.
        - Optional OTP challenge for MFA-enabled accounts.

        Permissions: Public
        Rate limit: 20 requests / 15 minutes per IP
      operationId: loginCustomer
      security: []
      x-permissions: [public]
      x-rateLimit:
        scope: ip
        burst: 20
        window: 15m
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            examples:
              default:
                value:
                  email: jordan.parker@atlasbank.com
                  password: StrongPass!234
                  otpCode: '123456'
      responses:
        '200':
          description: Authentication successful.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AuthSession'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '423':
          $ref: '#/components/responses/Locked'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Refresh access token
      description: |
        Issues a new short-lived access token using a refresh token.

        Validation:
        - Refresh token must be active and not revoked.

        Permissions: Public with refresh token
        Rate limit: 60 requests / hour per session
      operationId: refreshAuthToken
      security:
        - refreshTokenCookie: []
      x-permissions:
        - refresh_token
      x-rateLimit:
        scope: session
        burst: 60
        window: 1h
      requestBody:
        required: false
      responses:
        '200':
          description: Access token refreshed.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/TokenRefreshResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/logout:
    post:
      tags: [Authentication]
      summary: End active session
      description: |
        Revokes the current session and invalidates tokens.

        Permissions: customer:self
        Rate limit: 120 requests / hour per user
      operationId: logoutCustomer
      x-permissions:
        - customer:self
      x-rateLimit:
        scope: user
        burst: 120
        window: 1h
      responses:
        '200':
          description: Session ended.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
              examples:
                ok:
                  value:
                    success: true
                    data:
                      message: Session revoked.
                    meta:
                      traceId: trc_1002
                      timestamp: '2026-07-09T13:08:00Z'
                    error: null
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/password/forgot:
    post:
      tags: [Authentication]
      summary: Start password reset
      description: |
        Initiates a password reset flow and sends reset instructions.

        Permissions: Public
        Rate limit: 5 requests / hour per email
      operationId: initiatePasswordReset
      security: []
      x-permissions: [public]
      x-rateLimit:
        scope: subject
        burst: 5
        window: 1h
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '202':
          description: Reset request accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '429':
          $ref: '#/components/responses/TooManyRequests'
  /auth/password/reset:
    post:
      tags: [Authentication]
      summary: Complete password reset
      description: |
        Resets account password using a reset token.

        Validation:
        - Reset token validity and expiration.
        - New password policy validation.

        Permissions: Public with valid reset token
        Rate limit: 10 requests / hour per token
      operationId: completePasswordReset
      security: []
      x-permissions: [public]
      x-rateLimit:
        scope: token
        burst: 10
        window: 1h
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/sessions:
    get:
      tags: [Authentication]
      summary: List active sessions
      description: |
        Returns sessions for the authenticated customer.

        Supports filtering by status and ordering by last activity.

        Permissions: customer:self
        Rate limit: 300 requests / hour per user
      operationId: listSessions
      x-permissions: [customer:self]
      x-rateLimit:
        scope: user
        burst: 300
        window: 1h
      parameters:
        - $ref: '#/components/parameters/Offset'
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Sort'
        - $ref: '#/components/parameters/Order'
        - $ref: '#/components/parameters/Search'
      responses:
        '200':
          description: Session list returned.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiEnvelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Session'
                      meta:
                        $ref: '#/components/schemas/OffsetMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /auth/sessions/{sessionId}:
    delete:
      tags: [Authentication]
      summary: Revoke a specific session
      description: |
        Revokes a single session by identifier.

        Permissions: customer:self
        Rate limit: 120 requests / hour per user
      operationId: revokeSession
      x-permissions: [customer:self]
      x-rateLimit:
        scope: user
        burst: 120
        window: 1h
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: Session revoked.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token for authenticated customer/admin APIs.
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: Refresh token cookie used for token refresh endpoint.
    sessionCookie:
      type: apiKey
      in: cookie
      name: atlas_session
      description: Session cookie for session-based flows.
    oauth2:
      type: oauth2
      description: Future OAuth 2.1 support for external integrations.
      flows:
        authorizationCode:
          authorizationUrl: https://auth.atlasbank.com/oauth2/authorize
          tokenUrl: https://auth.atlasbank.com/oauth2/token
          scopes:
            atlas.read: Read access to Atlas resources.
            atlas.write: Write access to Atlas resources.
  parameters:
...
    Offset:
      name: offset
      in: query
      description: Offset for offset pagination.
      schema:
        type: integer
        minimum: 0
        default: 0
    Limit:
      name: limit
      in: query
      description: Maximum items to return.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 25
    Sort:
      name: sort
      in: query
      description: Field name used for sorting.
      schema:
        type: string
    Order:
      name: order
      in: query
      description: Sort direction.
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    Search:
      name: search
      in: query
      description: Full-text search term.
      schema:
        type: string
        minLength: 1
        maxLength: 120
...
    SessionId:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    Unauthorized:
      description: Authentication failed or missing credentials.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    NotFound:
      description: Requested resource not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: Resource conflict or incompatible state.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Locked:
      description: Account temporarily locked.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    InternalServerError:
      description: Unexpected server failure.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    ApiEnvelope:
      type: object
      required: [success, data, meta, error]
      properties:
        success:
          type: boolean
        data:
          nullable: true
        meta:
          $ref: '#/components/schemas/ResponseMeta'
        error:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/ErrorObject'
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/ApiEnvelope'
        - type: object
          properties:
            success:
              const: false
            data:
              type: 'null'
            error:
              $ref: '#/components/schemas/ErrorObject'
    ResponseMeta:
      type: object
      required: [traceId, timestamp]
      properties:
        traceId:
          type: string
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
        pagination:
          $ref: '#/components/schemas/PaginationMeta'
        rateLimit:
          $ref: '#/components/schemas/RateLimitMeta'
    PaginationMeta:
      type: object
      properties:
        mode:
          type: string
          enum: [cursor, offset]
        cursor:
          type: object
          properties:
            next:
              type: string
            previous:
              type: string
            hasMore:
              type: boolean
        offset:
          type: object
          properties:
            total:
              type: integer
            offset:
              type: integer
            limit:
              type: integer
    RateLimitMeta:
      type: object
      properties:
        limit:
          type: integer
        remaining:
          type: integer
        resetAt:
          type: string
          format: date-time
    ErrorObject:
      type: object
      required: [type, code, message]
      properties:
        type:
          type: string
          enum:
            - validation_error
            - authentication_error
            - authorization_error
            - business_rule_error
            - rate_limit_error
            - server_error
        code:
          type: string
        message:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              reason:
                type: string
        traceId:
          type: string
    OffsetMeta:
      allOf:
        - $ref: '#/components/schemas/ResponseMeta'
        - type: object
          properties:
            pagination:
              type: object
              properties:
                mode:
                  const: offset
                offset:
                  type: object
                  required: [total, offset, limit]
                  properties:
                    total:
                      type: integer
                    offset:
                      type: integer
                    limit:
                      type: integer
...
    RegisterRequest:
      type: object
      required: [email, password, firstName, lastName, termsAcceptedAt, privacyAcceptedAt]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
          maxLength: 128
        firstName:
          type: string
          minLength: 1
          maxLength: 100
        lastName:
          type: string
          minLength: 1
          maxLength: 100
        termsAcceptedAt:
          type: string
          format: date-time
        privacyAcceptedAt:
          type: string
          format: date-time
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
        otpCode:
          type: string
          pattern: '^[0-9]{6}$'
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    ResetPasswordRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
        newPassword:
          type: string
          minLength: 12
          maxLength: 128
    AuthSession:
      type: object
      required: [accessToken, refreshToken, tokenType, expiresIn, sessionId]
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
        sessionId:
          type: string
    TokenRefreshResponse:
      type: object
      required: [accessToken, tokenType, expiresIn]
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          enum: [Bearer]
        expiresIn:
          type: integer
    Session:
      type: object
      required: [id, device, ipAddress, createdAt, expiresAt, isCurrent]
      properties:
        id:
          type: string
          format: uuid
        device:
          type: string
        ipAddress:
          type: string
        userAgent:
          type: string
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        isCurrent:
          type: boolean
...